django-ros-turtlesim-server/requirements.txt
Django==3.0.3
twilio==6.35.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
djangorestframework==3.11.0
django-cors-headers==3.2.0
django-cors-middleware==1.5.0
roslibpy==0.7.1
drf-jwt==1.13.4
pyjwt==1.7.1
requests==2.22.0
schema==0.7.1
python-jose
social-auth-app-django
celery==4.4.0
redis==3.3.11
channels==2.4.0
channels_redis==2.4.2
